+++ projects/ngx-libs/src/lib/ngx-jsoneditor/classes/NGX_JSONEDITOR_MODE.enum.ts
@@ -1,0 +1,12 @@
+/** jsoneditor支持的mode
+ *
+ *
+ * @export
+ */
+export enum NGX_JSONEDITOR_MODE {
+    CODE = 'code',
+    TREE = 'tree',
+    TEXT = 'text',
+    VIEW = 'view',
+    FORM = 'form',
+}