+++ projects/ngx-libs/src/lib/ngx-jsoneditor/classes/index.ts
@@ -1,0 +1,3 @@
+export { NGX_JSONEDITOR_MODE } from "./NGX_JSONEDITOR_MODE.enum";
+export { NgxJsoneditorMode } from './ngx-jsoneditor-mode.type';
+export { NgxJsoneditorOptions } from './ngx-jsoneditor-options.class';