+++ projects/ngx-libs/src/lib/ngx-jsoneditor/classes/ngx-jsoneditor-options.class.ts
@@ -1,0 +1,11 @@
+import { NgxJsoneditorMode } from "./ngx-jsoneditor-mode.type";
+/** jsoneditor的配置信息
+ *
+ *
+ * @export
+ */
+export class NgxJsoneditorOptions {
+    public mode?: NgxJsoneditorMode;
+    public search?: boolean;
+    public onChange?: () => void;
+}