+++ projects/ngx-libs/src/lib/ngx-jsoneditor/ngx-jsoneditor.module.ts
@@ -1,0 +1,9 @@
+import { NgModule } from '@angular/core';
+import { NgxJsoneditorComponent } from './ngx-jsoneditor.component';
+
+@NgModule({
+  imports: [],
+  declarations: [NgxJsoneditorComponent],
+  exports: [NgxJsoneditorComponent]
+})
+export class NgxJsoneditorModule { }