+++ projects/ngx-libs/src/lib/ngx-perfect-scrollbar/index.ts
@@ -1,5 +1,3 @@
-import { from } from 'rxjs';
-
 export { NgxPerfectScrollbarModule } from './ngx-perfect-scrollbar.module';
 export { NgxPerfectScrollbarComponent } from './ngx-perfect-scrollbar.component';
 export { NgxPerfectScrollbarService } from './ngx-perfect-scrollbar.service';