+++ projects/ngx-libs/src/lib/ngx-prismjs/ngx-prismjs.module.ts
@@ -1,16 +1,9 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { NgxPrismjsDirective } from './ngx-prismjs.directive';
 
 @NgModule({
-   imports: [
-      CommonModule
-   ],
-   declarations: [
-      NgxPrismjsDirective
-   ],
-   exports: [
-      NgxPrismjsDirective
-   ]
+   imports: [],
+   declarations: [NgxPrismjsDirective],
+   exports: [NgxPrismjsDirective]
 })
 export class NgxPrismjsModule { }