+++ src/app/modules/jsoneditor-demo/jsoneditor-demo.routing.ts
@@ -1,0 +1,15 @@
+import { Routes, RouterModule } from '@angular/router';
+import { NgModule } from '@angular/core';
+import { JsoneditorDemoComponent } from './jsoneditor-demo.component';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: JsoneditorDemoComponent
+  },
+];
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class JsoneditorDemoRoutesModule { }